Error Handling FinalProject_AddressBasedHashTable.cpp
This edit added general error handling to process improper user inputs and errors. Specifically, the user menu in the auctionMenu() function was changed to a switch function for better readability from ifelse statements. Additionally, using cin.clear() and cin.ignore() lines error handling for noninteger inputs on the menu itself as well as for placing a bid integer value was added.